refactor: rename safeZoneTopBar directory to layout
- Rename safeZoneTopBar to layout directory to have a shorter name and ensure consistency
- Rename layout package files and rename references of all files that use those files
